outsourced evomap and fixed bugs introduced via PR
