+++ pokete_classes/poke.py
@@ -10,7 +10,7 @@ from pokete_general_use_fns import liner
 from .attack_actions import AttackActions
 from .attack import Attack
 from .health_bar import HealthBar
-from .fightmap import EvoMap
+from .evomap import EvoMap
 from .color import Color
 from .moves import Moves
 from .types import types
